requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==1.7.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
